requirements.txt
authlib==1.5.2
Flask==3.1.1
gdown==5.2.0
numpy==2.3.0
opencv_python==4.11.0.86
piexif==1.1.3
Pillow==11.2.1
PyJWT==2.10.1
PyJWT==2.10.1
python-dotenv==1.1.0
Requests==2.32.4
supervision==0.25.1